requirements.txt
python-dotenv
openai
genanki
gTTS
